move Point and Interpolate to database crate
