+++ database/src/lib.rs
@@ -9,6 +9,7 @@ use std::ops::{Add, Sub};
 use std::sync::Arc;
 use std::time::Duration;
 
+pub mod interpolate;
 pub mod pool;
 pub mod selector;
