+++ site/src/lib.rs
@@ -10,7 +10,6 @@ pub mod server;
 mod average;
 mod benchmark_metadata;
 mod comparison;
-mod interpolate;
 mod request_handlers;
 mod resources;
 mod selector;
